Refactorng script for download articles.
